# packages/docs/openapi/data-engine.yaml
openapi: 3.0.0
info:
    title: Data Engine API
    description: |
        API for managing structured data resources including domains, URLs, calls, tags, documents, tickets, and chats.

        This API provides comprehensive CRUD operations for managing data across multiple domains with support for filtering, pagination, and advanced search capabilities.
    version: 1.0.0
    contact:
        name: AnswerAgentAI Support
        url: https://theanswer.ai
        email: support@theanswer.ai
servers:
    - url: https://prod.studio.theanswer.ai/api/v1/data-engine
      description: Production server
    - url: http://localhost:4000/api/v1/data-engine
      description: Local development server
...
paths:
...
    /domains:
        post:
            summary: Create a new domain
            description: Create a new domain record with metadata and custom data
            operationId: createDomain
            tags:
                - domains
            security:
                - bearerAuth: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/DomainInput'
            responses:
                '201':
                    description: Domain created successfully
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Domain'
                '400':
                    description: Invalid input provided
                '401':
                    description: Unauthorized - Invalid API key
                '500':
                    description: Internal server error
        get:
            summary: List all domains
            description: Retrieve domains with filtering and pagination
            operationId: listDomains
            tags:
                - domains
            security:
                - bearerAuth: []
            parameters:
                - $ref: '#/components/parameters/PageParam'
                - $ref: '#/components/parameters/PageSizeParam'
                - $ref: '#/components/parameters/SearchTermParam'
                - name: includeTags
                  in: query
                  description: Include domains with tags (JSON string)
                  schema:
                      type: string
                      example: '{"tags": ["tag-slug"], "logic": "OR"}'
                - name: excludeTags
                  in: query
                  description: Exclude domains with tags (JSON string)
                  schema:
                      type: string
                      example: '{"tags": ["tag-slug"], "logic": "OR"}'
                - name: isValid
                  in: query
                  description: Filter by validity status
                  schema:
                      type: boolean
                - name: hasAnalysis
                  in: query
                  description: Filter by analysis status
                  schema:
                      type: string
                      enum: [all, has, not]
                - name: customFilters
                  in: query
                  description: Custom field filters (JSON string)
                  schema:
                      type: string
                      example: '{"field_key": "value"}'
            responses:
                '200':
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DomainListResponse'
                '401':
                    description: Unauthorized
                '500':
                    description: Internal server error
    /domains/{id}:
        get:
            summary: Get domain by ID
            description: Retrieve a single domain by ID
            operationId: getDomainById
            tags:
                - domains
            security:
                - bearerAuth: []
            parameters:
                - $ref: '#/components/parameters/IdParam'
            responses:
                '200':
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Domain'
                '404':
                    description: Domain not found
                '401':
                    description: Unauthorized
        put:
            summary: Update domain
            description: Update an existing domain
            operationId: updateDomain
            tags:
                - domains
            security:
                - bearerAuth: []
            parameters:
                - $ref: '#/components/parameters/IdParam'
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/DomainUpdate'
            responses:
                '200':
                    description: Domain updated successfully
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Domain'
                '400':
                    description: Invalid input
                '404':
                    description: Domain not found
                '401':
                    description: Unauthorized
        delete:
            summary: Delete domain
            description: Delete a domain by ID
            operationId: deleteDomain
            tags:
                - domains
            security:
                - bearerAuth: []
            parameters:
                - $ref: '#/components/parameters/IdParam'
            responses:
                '200':
                    description: Domain deleted successfully
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DeleteResponse'
                '404':
                    description: Domain not found
                '401':
                    description: Unauthorized
    # ==================== URLS ====================
...
    /documents:
        post:
            summary: Create a new document
            description: Create a new document with optional embedding
            operationId: createDocument
            tags:
                - documents
            security:
                - bearerAuth: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/DocumentInput'
            responses:
                '201':
                    description: Document created successfully
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Document'
        get:
            summary: List all documents
            operationId: listDocuments
            tags:
                - documents
            security:
                - bearerAuth: []
            parameters:
                - $ref: '#/components/parameters/PageParam'
                - $ref: '#/components/parameters/PageSizeParam'
                - name: store_id
                  in: query
                  schema:
                      type: string
                      format: uuid
            responses:
                '200':
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DocumentListResponse'
    /documents/search:
        post:
            summary: Search documents
            description: Search documents using vector similarity
            operationId: searchDocuments
            tags:
                - documents
            security:
                - bearerAuth: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/DocumentSearchInput'
            responses:
                '200':
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DocumentSearchResponse'
    /documents/{id}:
        get:
            summary: Get document by ID
            operationId: getDocumentById
            tags:
                - documents
            security:
                - bearerAuth: []
            parameters:
                - $ref: '#/components/parameters/IdParam'
            responses:
                '200':
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Document'
        put:
            summary: Update document
            operationId: updateDocument
            tags:
                - documents
            security:
                - bearerAuth: []
            parameters:
                - $ref: '#/components/parameters/IdParam'
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/DocumentUpdate'
            responses:
                '200':
                    description: Document updated successfully
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Document'
        delete:
            summary: Delete document
            operationId: deleteDocument
            tags:
                - documents
            security:
                - bearerAuth: []
            parameters:
                - $ref: '#/components/parameters/IdParam'
            responses:
                '200':
                    description: Document deleted successfully
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DeleteResponse'
    # ==================== TICKETS ====================
...
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: API Key
            description: |
                API key authentication. Include your API key in the Authorization header:
                `Authorization: Bearer YOUR_API_KEY`
    parameters:
        PageParam:
            name: page
            in: query
            description: Page number (0-indexed)
            schema:
                type: integer
                default: 0
                minimum: 0
        PageSizeParam:
            name: pageSize
            in: query
            description: Items per page
            schema:
                type: integer
                default: 10
                minimum: 1
                maximum: 100
        SearchTermParam:
            name: searchTerm
            in: query
            description: Search term for filtering results
            schema:
                type: string
        IdParam:
            name: id
            in: path
            required: true
            description: Resource ID (UUID)
            schema:
                type: string
                format: uuid
    schemas:
...
        Metadata:
            type: object
            description: Metadata object for tracking source system and user information
            properties:
                source_system:
                    type: string
                    example: theanswer
                source_organization_id:
                    type: string
                    format: uuid
                source_user_id:
                    type: string
                    format: uuid
                created_by:
                    type: string
                    format: email
                last_updated_by:
                    type: string
                    format: email
                last_updated_from:
                    type: string
                    example: theanswer
        DeleteResponse:
            type: object
            properties:
                success:
                    type: boolean
                    example: true
                message:
                    type: string
                    example: Resource deleted successfully
        PaginationInfo:
            type: object
            properties:
                total:
                    type: integer
                    example: 100
                page:
                    type: integer
                    example: 0
                pageSize:
                    type: integer
                    example: 10
        # Domain Schemas
        DomainInput:
            type: object
            required:
                - domain_name
            properties:
                domain_name:
                    type: string
                    example: example.com
                is_valid:
                    type: boolean
                    example: true
                meta_title:
                    type: string
                    example: Example Domain
                meta_description:
                    type: string
                    example: Example domain description
                custom_data:
                    type: object
                    additionalProperties: true
                    example:
                        source: manual
                        notes: Added manually
                metadata:
                    $ref: '#/components/schemas/Metadata'
        Domain:
            allOf:
                - $ref: '#/components/schemas/DomainInput'
                - type: object
                  properties:
                      id:
                          type: string
                          format: uuid
                      created_at:
                          type: string
                          format: date-time
                      updated_at:
                          type: string
                          format: date-time
                      last_analyzed_at:
                          type: string
                          format: date-time
                          nullable: true
                      ai_analysis:
                          type: object
                          additionalProperties: true
        DomainUpdate:
            type: object
            properties:
                meta_title:
                    type: string
                is_valid:
                    type: boolean
                custom_data:
                    type: object
                    additionalProperties: true
                metadata:
                    $ref: '#/components/schemas/Metadata'
        DomainListResponse:
            allOf:
                - $ref: '#/components/schemas/PaginationInfo'
                - type: object
                  properties:
                      domains:
                          type: array
                          items:
                              $ref: '#/components/schemas/Domain'
        # URL Schemas
...
        DocumentInput:
            type: object
            required:
                - content
            properties:
                content:
                    type: string
                    example: Document content text...
                embedding:
                    type: array
                    items:
                        type: number
                    example: [0.1, 0.2, 0.3]
                metadata:
                    type: object
                    additionalProperties: true
                    example:
                        title: Document Title
                        source: upload
                store_id:
                    type: string
                    format: uuid
        Document:
            allOf:
                - $ref: '#/components/schemas/DocumentInput'
                - type: object
                  properties:
                      id:
                          type: string
                          format: uuid
                      created_at:
                          type: string
                          format: date-time
                      updated_at:
                          type: string
                          format: date-time
        DocumentUpdate:
            type: object
            properties:
                content:
                    type: string
                metadata:
                    type: object
                    additionalProperties: true
        DocumentSearchInput:
            type: object
            required:
                - query_embedding
            properties:
                query_embedding:
                    type: array
                    items:
                        type: number
                    description: Vector embedding for similarity search
                    example: [0.1, 0.2, 0.3]
                match_threshold:
                    type: number
                    format: float
                    default: 0.7
                    description: Minimum similarity threshold
                match_count:
                    type: integer
                    default: 10
                    description: Maximum number of results
                store_id:
                    type: string
                    format: uuid
                    description: Optional store ID to filter by
        DocumentSearchResponse:
            type: object
            properties:
                documents:
                    type: array
                    items:
                        $ref: '#/components/schemas/Document'
                count:
                    type: integer
                    example: 10
        DocumentListResponse:
            allOf:
                - $ref: '#/components/schemas/PaginationInfo'
                - type: object
                  properties:
                      documents:
                          type: array
                          items:
                              $ref: '#/components/schemas/Document'
        # Ticket Schemas
...